Move associados to accounts
