--- associados/apps.py
@@ -1,8 +1,0 @@
-from django.apps import AppConfig
-
-
-class AssociadosConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "associados"
-    verbose_name = "Associados"
-
