+++ templates/base.html
@@ -98,7 +98,7 @@
         <a href="/" class="flex items-center gap-x-2 hover:text-primary transition">
           <i class="fa-solid fa-gauge"></i> {% trans "Dashboard" %}
         </a>
-        <a href="{% url 'associados:lista' %}" class="flex items-center gap-x-2 hover:text-primary transition">
+        <a href="{% url 'accounts:associados_lista' %}" class="flex items-center gap-x-2 hover:text-primary transition">
           <i class="fa-solid fa-users"></i> {% trans "Associados" %}
         </a>
         <a href="{% url 'empresas:lista' %}" class="flex items-center gap-x-2 hover:text-primary transition">
